requirements.txt
Flask==2.3.2
face_recognition==1.3.0
Pillow==10.0.0
# sqlite3==
# numpy==
